randstruct: Simplified field order test comparison
The previous implementation required an additional counter since using
just an iterator was insufficient. That is, if the iterator is
incremented past the end it still passed the test, which is *wrong*.

This implementation makes it easier to make assertions as well as see
the difference between the assertions when they fail.